refactor: change the way in which we create reducers and the store
Use the recommended approach in redux.js.org to setup the store and
reducers.

It is better to have all the reducers preloaded and just pass an initial
state to the render method.

Added some utils and improved some of the tests/code.

I've also remove the snapshot test as it is quite unstable. It uses
`useId` internally which might change between re-renders. Snapshot
testing provides little value anyway.